+++ DataGenerator/Program.cs
@@ -385,8 +385,8 @@ namespace DataGenerator
                         {
                             //SkillInfo.AddUnreleasedMoveData();
                             //SkillInfo.AddMoveData();
-                            SkillInfo.AddMoveData(533);
-                            //SkillInfo.AddMoveDataToAnims(364, 566, 467);
+                            //SkillInfo.AddMoveData(6);
+                            SkillInfo.AddMoveDataToAnims(15,75,163,210,314,332,348,400,403,404,427,440,533,534,669,781,830,845,860,869,875,891,895);
                         }
 
                         if ((dump & DataManager.DataType.Intrinsic) != DataManager.DataType.None)