+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # 0.7.16 Changes #
 
-* Fixed the learnsets for Pokemon introduced in Pokemon Legends Arceus
+* Fixed the learnset for Overqwil
+* Fixed Sharpness ability
 * Fixed an issue where grass couldnt be cut
 * Fixed an issue where wild pokemon would grab the treasure of Castaway Cave
 * Fixed an issue where Amber Tear would stack from user and target.
 * Fixed issues where Corrosion and Infiltrator would affect the user.
 * Fixed Bug Bite/Pluck's interaction with Ban Seed
+* Rest PP 10 -> 8
 * Fixed error occurring when warping off of water due to stepping on a gust trap due to warping off of water due to not being able to walk on water
 * Dev: Ability to add decorations in dungeon maps
 * Dev: More friendly strings for editors